Separated validation into an interface. Moved to the common package.
