+++ src/Titon/Common/Validator/CoreValidator.hh
@@ -1,0 +1,17 @@
+<?hh // strict
+/**
+ * @copyright   2010-2015, The Titon Project
+ * @license     http://opensource.org/licenses/bsd-license.php
+ * @link        http://titon.io
+ */
+
+namespace Titon\Common\Validator;
+
+/**
+ * Core class for providing validation.
+ *
+ * @package Titon\Common\Validator
+ */
+class CoreValidator extends AbstractValidator {
+
+}
